server: test proxy server
Updates the proxy server inline tests to test the server rather than the
HTTP proxy handler.
